xfs: throttle inodegc queuing on backlog
Track the number of inodes in each AG that are queued for inactivation,
then use that information to decide if we're going to make threads that
has queued an inode for inactivation wait for the background thread.
The purpose of this high water mark is to establish a maximum bound on
the backlog of work that can accumulate on a non-frozen filesystem.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>