xfs: reduce inactivation delay when free space is tight
Now that we have made the inactivation of unlinked inodes a background
task to increase the throughput of file deletions, we need to be a
little more careful about how long of a delay we can tolerate.

On a mostly empty filesystem, the risk of the allocator making poor
decisions due to fragmentation of the free space on account a lengthy
delay in background updates is minimal because there's plenty of space.
However, if free space is tight, we want to deallocate unlinked inodes
as quickly as possible to avoid fallocate ENOSPC and to give the
allocator the best shot at optimal allocations for new writes.

Therefore, use the same free space thresholds that we use to limit
preallocation to scale down the delay between an AG being tagged for
needing inodgc work and the inodegc worker being executed.  This follows
the same principle that XFS becomes less aggressive about allocations
(and more precise about accounting) when nearing full.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>