xfs: Fix incorrect AG locking order during symbolic link creation
During the creation of a symbolic link, inode allocation causes the
corresponding AG's AGI to be locked. Later, when allocating space for symbolic
link's blocks, the task could wrap around the AGs and try to lock the AGF of
an AG whose agno is less than that of the AGI that was prevously locked. This
breaks the AG locking order rule.

This commit fixes the incorrect locking order by rolling the transaction
before allocating blocks for the symbolic link.

Signed-off-by: Chandan Babu R <chandanbabu@kernel.org>